refactor: split LobbyCommands.cs (1,452→493 lines) — extract command handlers to partial class
Extracted all 24 static command handler methods into
LobbyCommands.Handlers.cs using a C# partial class. Zero
behavioral change.

LobbyCommands.cs (493 lines): class declaration, Fluent constants,
  constructor, ValidateSlotCommand, ValidateCommand,
  InterpretCommand (dispatch), CheckAutoStart, server lifecycle
  (InitializeMapPool, ServerStarted, LoadMapSettings, ClientJoined,
  ServerEmpty)
LobbyCommands.Handlers.cs (986 lines): 24 command handlers —
  State, StartGame, Slot, AllowSpectators, Specate, SlotClose,
  SlotOpen, SlotBot, Map, Option, ResetOptions, AssignTeams, Kick,
  VoteKick, MakeAdmin, MakeSpectator, Name, Faction, Team,
  Handicap, ClearPlayerSpawn, Spawn, PlayerColor, SyncLobby
